Minor important changes: (2018-08-11) Update source code and binaries and script files
Remove the meaningless C++ language identifier, and all the text content of the function printed to the screen has been corrected and added to the string content.
Some files or folders have been renamed.
