+++ [Encrypt_Or_Decrypt_File]-Source-Code/Beta-Version/StdAfx.h
@@ -16,7 +16,7 @@
 #define DLLEXPORT __declspec(dllexport)
 #define	DLLIMPORT __declspec(dllimport)
 
-#include "stdio.h"
+//#include <stdio.h>
 #include <windows.h>
 #include <winsock2.h>
 #include <ws2tcpip.h>
