+++ [Encrypt_Or_Decrypt_File]-Source-Code/Beta-Version3/EODF3_By_TDOM.hpp
@@ -228,7 +228,7 @@
                  }
               fwrite(buffer, 16, FileByteData, FilePointerTarget); //将buffer中的数据写入文件
             }
-        };
+        }
 
 /***********************************************************************************************************************************************************/
 
@@ -473,7 +473,7 @@
                  }
               fwrite(buffer, 16, FileByteData, FilePointerTarget); //将buffer中的数据写入文件
             }
-         };
+         }
 
 /***********************************************************************************************************************************************************/
